Phase 6d: Extract schedule loader into useScheduleLoader hook
Extract getSchedule (~947 lines) and its 5 helper functions (~236 lines)
from App.jsx into a dedicated useScheduleLoader hook, following the
established store-slice pattern. This is the single largest extraction
of the entire refactoring effort.

New files created:
- src/hooks/useScheduleLoader.js (1,008 lines): Contains the full
  getSchedule function and fetchTBAMatches helper. Receives ~30 deps
  (state reads, setters, refs, external functions, concrete store
  functions) and returns { getSchedule }.
- src/utils/scheduleConformance.js (~260 lines): Pure data transformer
  functions extracted from App.jsx — conformCheesyArenaMatch,
  conformCheesyArenaScores, conformCFTCOfflineScheduleMatch,
  conformFTCOfflineScores, and winner(match, ftcMode).

Modified files:
- src/contexts/EventStoreProvider.jsx: Added useScheduleLoader as the
  4th slice (team → rankings → schedule → matchNav). Schedule slice is
  created BEFORE matchNav slice so that matchNavDeps can be composed
  with scheduleSlice.getSchedule. Added getSchedule to storeRef.current
  and allActions.
- src/App.jsx: Removed getSchedule function (947 lines), fetchTBAMatches
  (16 lines), and 5 conformer helpers (236 lines). Built scheduleDeps
  object with ~30 properties. Switched 4 internal callers to
  eventStoreRef.current.getSchedule(). Removed unused imports
  (ftcHybridMatchTeams, ftcPlayoffSchedule, frcPlayoffSchedule,
  prunePlayoffReserveSetsAfterPostedMatches).

Key design decisions:
- winner(match, ftcMode) extracted with explicit ftcMode parameter
  instead of closing over it (was a nested closure in getSchedule)
- Concrete store functions (getTeamList, getAlliances, getRanks) passed
  via deps instead of passing the entire eventStoreRef
- training object passed via deps instead of re-importing/re-cloning
- Internal remapStringToNumber defined locally in hook (reads
  deps.teamRemappings) — the useCallback version stays in App.jsx
  for context consumers
- ftcBaseURL defined as module-level constant in the hook file
- Slice composition in EventStoreProvider: composedScheduleDeps adds
  getTeamList/getAlliances/getRanks from earlier slices;
  composedMatchNavDeps overrides getSchedule from scheduleSlice

App.jsx: 4,348 → 3,192 lines (1,156 lines removed, -27%)
